+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All Notable changes for the Laravel 5 WooCommerce REST API Client  will be documented in this file
 
+## 2.3.0
+- Add pagination functionality.
+Now these functions are available
+    - totalResults()
+    - firstPage()
+    - lastPage()
+    - currentPage()
+    - totalPages()
+    - previousPage()
+    - nextPage()
+    - hasPreviousPage()
+    - hasNextPage()
+    - hasNotPreviousPage()
+    - hasNotNextPage()
+- Get the Request & Response (Headers) from the last request
+    - getRequest()
+    - getResponse()
+
 ## 2.2.0
 - Laravel 5.4 compatibility